BLOGPROJECT/requirements.txt
alembic==1.11.1
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.4
dnspython==2.4.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
